added support for grouped and ungrouped slots
I was doing some silly thing where I was checking
the size of vector<FunctionType> slots; and returning
a default-constructed ReturnType if it was empty.
First of all, slots.empty() is preferable to
slots.size() == 0. Then I had a for-loop to iterate
over every single element, and if the next() element
== slots.end(), then I'd return it. But the compiler
shouldn't be able to optimize away returning the value
of the function anyways, so I really don't know what I
thought I was doing. In my defence, it was wine-
influenced coding around 2 AM, if memory serves.